Move components only used by beatmap set overlay local to namespace
