+++ osu.Game/Overlays/BeatmapSet/Details.cs
@@ -11,7 +11,6 @@ using osu.Game.Beatmaps;
 using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Overlays.BeatmapSet.Buttons;
 using osu.Game.Rulesets;
-using osu.Game.Screens.Select.Details;
 using osuTK;
 
 namespace osu.Game.Overlays.BeatmapSet
