+++ osu.Game/Overlays/BeatmapSet/FailRetryGraph.cs
@@ -4,15 +4,15 @@
 #nullable disable
 
 using System;
+using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Game.Beatmaps;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
-using System.Linq;
-using osu.Game.Beatmaps;
 
-namespace osu.Game.Screens.Select.Details
+namespace osu.Game.Overlays.BeatmapSet
 {
     public partial class FailRetryGraph : Container
     {
